requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
requests>=2.28.0
PyYAML>=6.0

# Database
pymysql>=1.0.2
# or use mysqlclient if preferred

# Scheduling
schedule>=1.1.0

# Optional: For better logging
python-json-logger>=2.0.4

# Optional: For service management (Windows)
# pywin32>=305  # Uncomment if using Windows Service

# Development
pytest>=7.2.0
black>=22.0.0
flake8>=5.0.0